Adds ability to set custom request handlers.
