+++ stripe/__init__.py
@@ -4,15 +4,12 @@
 
 ## Imports
 import logging
-import os
 import platform
-import sys
 import urllib
-import urlparse
-import textwrap
-import time
-import datetime
-import types
+
+from . import protocols
+from . import encoders
+from . import exceptions as stripe_exceptions
 
 # Use cStringIO if it's available.  Otherwise, StringIO is fine.
 try:
@@ -20,54 +17,8 @@ try:
 except ImportError:
   import StringIO
 
-# - Requests is the preferred HTTP library
-# - Google App Engine has urlfetch
-# - Use Pycurl if it's there (at least it verifies SSL certs)
-# - Fall back to urllib2 with a warning if needed
-_httplib = None
 
-try:
-  from google.appengine.api import urlfetch
-  _httplib = 'urlfetch'
-except ImportError:
-  pass
 
-if not _httplib:
-  try:
-    import requests
-    _httplib = 'requests'
-  except ImportError:
-    pass
-
-  try:
-    # Require version 0.8.8, but don't want to depend on distutils
-    version = requests.__version__
-    major, minor, patch = [int(i) for i in version.split('.')]
-  except Exception:
-    # Probably some new-fangled version, so it should support verify
-    pass
-  else:
-    if (major, minor, patch) < (0, 8, 8):
-      print >>sys.stderr, 'Warning: the Stripe library requires that your Python "requests" library has a version no older than 0.8.8, but your "requests" library has version %s. Stripe will fall back to an alternate HTTP library, so everything should work, though we recommend upgrading your "requests" library. If you have any questions, please contact support@stripe.com. (HINT: running "pip install -U requests" should upgrade your requests library to the latest version.)' % (version, )
-      _httplib = None
-
-if not _httplib:
-  try:
-    import pycurl
-    _httplib = 'pycurl'
-  except ImportError:
-    pass
-
-if not _httplib:
-  try:
-    import urllib2
-    _httplib = 'urllib2'
-    print >>sys.stderr, "Warning: the Stripe library is falling back to urllib2 because neither requests nor pycurl are installed. urllib2's SSL implementation doesn't verify server certificates. For improved security, we suggest installing requests."
-  except ImportError:
-    pass
-
-if not _httplib:
-  raise ImportError("Stripe requires one of requests, pycurl, Google App Engine's urlfetch, or urllib2.  If you are on a platform where none of these libraries are available, please let us know at support@stripe.com.")
 
 from version import VERSION
 import importer
@@ -80,38 +31,7 @@ logger = logging.getLogger('stripe')
 api_key = None
 api_base = 'https://api.stripe.com'
 api_version = None
-verify_ssl_certs = True
 
-## Exceptions
-class StripeError(Exception):
-  def __init__(self, message=None, http_body=None, http_status=None, json_body=None):
-    super(StripeError, self).__init__(message)
-    self.http_body = http_body and http_body.decode('utf-8')
-    self.http_status = http_status
-    self.json_body = json_body
-
-class APIError(StripeError):
-  pass
-
-class APIConnectionError(StripeError):
-  pass
-
-class CardError(StripeError):
-  def __init__(self, message, param, code, http_body=None, http_status=None, json_body=None):
-    super(CardError, self).__init__(message, http_body, http_status, json_body)
-    self.param = param
-    self.code = code
-    self.http_body = http_body and http_body.decode('utf-8')
-    self.http_status = http_status
-    self.json_body = json_body
-
-class InvalidRequestError(StripeError):
-  def __init__(self, message, param, http_body=None, http_status=None, json_body=None):
-    super(InvalidRequestError, self).__init__(message, http_body, http_status, json_body)
-    self.param = param
-
-class AuthenticationError(StripeError):
-  pass
 
 def convert_to_stripe_object(resp, api_key):
   types = { 'charge' : Charge, 'customer' : Customer,
@@ -143,66 +63,6 @@ class APIRequestor(object):
     return '%s%s' % (api_base, url)
 
   @classmethod
-  def _utf8(cls, value):
-    if isinstance(value, unicode) and sys.version_info < (3, 0):
-      return value.encode('utf-8')
-    else:
-      return value
-
-  @classmethod
-  def encode_dict(cls, stk, key, dictvalue):
-    n = {}
-    for k, v in dictvalue.iteritems():
-      k = cls._utf8(k)
-      v = cls._utf8(v)
-      n["%s[%s]" % (key, k)] = v
-    stk.extend(cls._encode_inner(n))
-
-
-  @classmethod
-  def encode_list(cls, stk, key, listvalue):
-    for v in listvalue:
-      v = cls._utf8(v)
-      stk.append(("%s[]" % (key), v))
-
-  @classmethod
-  def encode_datetime(cls, stk, key, dttime):
-    utc_timestamp = int(time.mktime(dttime.timetuple()))
-    stk.append((key, utc_timestamp))
-
-  @classmethod
-  def encode_none(cls, stk, k, v):
-    pass # do not include None-valued params in request
-
-  @classmethod
-  def _encode_inner(cls, d):
-    """
-    We want post vars of form:
-    {'foo': 'bar', 'nested': {'a': 'b', 'c': 'd'}}
-    to become:
-    foo=bar&nested[a]=b&nested[c]=d
-    """
-    # special case value encoding
-    ENCODERS = {
-      list: cls.encode_list,
-      dict: cls.encode_dict,
-      datetime.datetime: cls.encode_datetime,
-      types.NoneType: cls.encode_none,
-    }
-
-    stk = []
-    for key, value in d.iteritems():
-      key = cls._utf8(key)
-      try:
-          encoder = ENCODERS[value.__class__]
-          encoder(stk, key, value)
-      except KeyError:
-        # don't need special encoding
-        value = cls._utf8(value)
-        stk.append((key, value))
-    return stk
-
-  @classmethod
   def _objects_to_ids(cls, d):
     if isinstance(d, APIResource):
       return d.id
@@ -214,20 +74,6 @@ class APIRequestor(object):
     else:
       return d
 
-  @classmethod
-  def encode(cls, d):
-    """
-    Internal: encode a string for url representation
-    """
-    return urllib.urlencode(cls._encode_inner(d))
-
-  @classmethod
-  def build_url(cls, url, params):
-    base_query = urlparse.urlparse(url).query
-    if base_query:
-      return '%s&%s' % (url, cls.encode(params))
-    else:
-      return '%s?%s' % (url, cls.encode(params))
 
   def request(self, meth, url, params={}):
     rbody, rcode, my_api_key = self.request_raw(meth, url, params)
@@ -238,16 +84,16 @@ class APIRequestor(object):
     try:
       error = resp['error']
     except (KeyError, TypeError):
-      raise APIError("Invalid response object from API: %r (HTTP response code was %d)" % (rbody, rcode), rbody, rcode, resp)
+      raise stripe_exceptions.APIError("Invalid response object from API: %r (HTTP response code was %d)" % (rbody, rcode), rbody, rcode, resp)
 
     if rcode in [400, 404]:
-      raise InvalidRequestError(error.get('message'), error.get('param'), rbody, rcode, resp)
+      raise stripe_exceptions.InvalidRequestError(error.get('message'), error.get('param'), rbody, rcode, resp)
     elif rcode == 401:
-      raise AuthenticationError(error.get('message'), rbody, rcode, resp)
+      raise stripe_exceptions.AuthenticationError(error.get('message'), rbody, rcode, resp)
     elif rcode == 402:
-      raise CardError(error.get('message'), error.get('param'), error.get('code'), rbody, rcode, resp)
+      raise stripe_exceptions.CardError(error.get('message'), error.get('param'), error.get('code'), rbody, rcode, resp)
     else:
-      raise APIError(error.get('message'), rbody, rcode, resp)
+      raise stripe_exceptions.APIError(error.get('message'), rbody, rcode, resp)
 
   def request_raw(self, meth, url, params={}):
     """
@@ -255,7 +101,7 @@ class APIRequestor(object):
     """
     my_api_key = self.api_key or api_key
     if my_api_key is None:
-      raise AuthenticationError('No API key provided. (HINT: set your API key using "stripe.api_key = <API-KEY>"). You can generate API keys from the Stripe web interface.  See https://stripe.com/api for details, or email support@stripe.com if you have any questions.')
+      raise stripe_exceptions.AuthenticationError('No API key provided. (HINT: set your API key using "stripe.api_key = <API-KEY>"). You can generate API keys from the Stripe web interface.  See https://stripe.com/api for details, or email support@stripe.com if you have any questions.')
 
     abs_url = self.api_url(url)
     params = params.copy()
@@ -265,7 +111,7 @@ class APIRequestor(object):
       'bindings_version' : VERSION,
       'lang' : 'python',
       'publisher' : 'stripe',
-      'httplib': _httplib,
+      'httplib': protocols.protocol,
       }
     for attr, func in [['lang_version', platform.python_version],
                        ['platform', platform.platform],
@@ -283,17 +129,11 @@ class APIRequestor(object):
       }
     if api_version is not None:
       headers['Stripe-Version'] = api_version
+    requestor = protocols.get_requestor()
+    if requestor is None:
+        raise stripe_exceptions.StripeError("Stripe Python library bug discovered: invalid httplib %s.  Please report to support@stripe.com" % (protocols.protocol, ))
+    rbody, rcode = requestor(meth, abs_url, headers, params)
 
-    if _httplib == 'requests':
-      rbody, rcode = self.requests_request(meth, abs_url, headers, params)
-    elif _httplib == 'pycurl':
-      rbody, rcode = self.pycurl_request(meth, abs_url, headers, params)
-    elif _httplib == 'urlfetch':
-      rbody, rcode = self.urlfetch_request(meth, abs_url, headers, params)
-    elif _httplib == 'urllib2':
-      rbody, rcode = self.urllib2_request(meth, abs_url, headers, params)
-    else:
-      raise StripeError("Stripe Python library bug discovered: invalid httplib %s.  Please report to support@stripe.com" % (_httplib, ))
     logger.info('API request to %s returned (response code, response body) of (%d, %r)' % (abs_url, rcode, rbody))
     return rbody, rcode, my_api_key
 
@@ -301,180 +141,10 @@ class APIRequestor(object):
     try:
       resp = json.loads(rbody.decode('utf-8'))
     except Exception:
-      raise APIError("Invalid response body from API: %s (HTTP response code was %d)" % (rbody, rcode), rbody, rcode)
+      raise stripe_exceptions.APIError("Invalid response body from API: %s (HTTP response code was %d)" % (rbody, rcode), rbody, rcode)
     if not (200 <= rcode < 300):
       self.handle_api_error(rbody, rcode, resp)
     return resp
-
-  def requests_request(self, meth, abs_url, headers, params):
-    meth = meth.lower()
-    if meth == 'get' or meth == 'delete':
-      if params:
-        abs_url = self.build_url(abs_url, params)
-      data = None
-    elif meth == 'post':
-      data = self.encode(params)
-    else:
-      raise APIConnectionError('Unrecognized HTTP method %r.  This may indicate a bug in the Stripe bindings.  Please contact support@stripe.com for assistance.' % (meth, ))
-
-    kwargs = {}
-    if verify_ssl_certs:
-      kwargs['verify'] = os.path.join(os.path.dirname(__file__), 'data/ca-certificates.crt')
-    else:
-      kwargs['verify'] = false
-
-    try:
-      try:
-        result = requests.request(meth, abs_url,
-                                  headers=headers, data=data, timeout=80,
-                                  **kwargs)
-      except TypeError, e:
-        raise TypeError('Warning: It looks like your installed version of the "requests" library is not compatible with Stripe\'s usage thereof. (HINT: The most likely cause is that your "requests" library is out of date. You can fix that by running "pip install -U requests".) The underlying error was: %s' %(e, ))
-
-      # This causes the content to actually be read, which could cause
-      # e.g. a socket timeout. TODO: The other fetch methods probably
-      # are succeptible to the same and should be updated.
-      content = result.content
-      status_code = result.status_code
-    except Exception, e:
-      # Would catch just requests.exceptions.RequestException, but can
-      # also raise ValueError, RuntimeError, etc.
-      self.handle_requests_error(e)
-    return content, status_code
-
-  def handle_requests_error(self, e):
-    if isinstance(e, requests.exceptions.RequestException):
-      msg = "Unexpected error communicating with Stripe.  If this problem persists, let us know at support@stripe.com."
-      err = "%s: %s" % (type(e).__name__, str(e))
-    else:
-      msg = "Unexpected error communicating with Stripe.  It looks like there's probably a configuration issue locally.  If this problem persists, let us know at support@stripe.com."
-      err = "A %s was raised" % (type(e).__name__, )
-      if str(e):
-        err += " with error message %s" % (str(e), )
-      else:
-        err += " with no error message"
-    msg = textwrap.fill(msg) + "\n\n(Network error: " + err + ")"
-    raise APIConnectionError(msg)
-
-  def pycurl_request(self, meth, abs_url, headers, params):
-    s = StringIO.StringIO()
-    curl = pycurl.Curl()
-
-    meth = meth.lower()
-    if meth == 'get':
-      curl.setopt(pycurl.HTTPGET, 1)
-      # TODO: maybe be a bit less manual here
-      if params:
-        abs_url = self.build_url(abs_url, params)
-    elif meth == 'post':
-      curl.setopt(pycurl.POST, 1)
-      curl.setopt(pycurl.POSTFIELDS, self.encode(params))
-    elif meth == 'delete':
-      curl.setopt(pycurl.CUSTOMREQUEST, 'DELETE')
-      if params:
-        abs_url = self.build_url(abs_url, params)
-    else:
-      raise APIConnectionError('Unrecognized HTTP method %r.  This may indicate a bug in the Stripe bindings.  Please contact support@stripe.com for assistance.' % (meth, ))
-
-    # pycurl doesn't like unicode URLs
-    abs_url = self._utf8(abs_url)
-    curl.setopt(pycurl.URL, abs_url)
-    curl.setopt(pycurl.WRITEFUNCTION, s.write)
-    curl.setopt(pycurl.NOSIGNAL, 1)
-    curl.setopt(pycurl.CONNECTTIMEOUT, 30)
-    curl.setopt(pycurl.TIMEOUT, 80)
-    curl.setopt(pycurl.HTTPHEADER, ['%s: %s' % (k, v) for k, v in headers.iteritems()])
-    if verify_ssl_certs:
-      curl.setopt(pycurl.CAINFO, os.path.join(os.path.dirname(__file__), 'data/ca-certificates.crt'))
-    else:
-      curl.setopt(pycurl.SSL_VERIFYHOST, False)
-
-    try:
-      curl.perform()
-    except pycurl.error, e:
-      self.handle_pycurl_error(e)
-    rbody = s.getvalue()
-    rcode = curl.getinfo(pycurl.RESPONSE_CODE)
-    return rbody, rcode
-
-  def handle_pycurl_error(self, e):
-    if e[0] in [pycurl.E_COULDNT_CONNECT,
-                pycurl.E_COULDNT_RESOLVE_HOST,
-                pycurl.E_OPERATION_TIMEOUTED]:
-      msg = "Could not connect to Stripe (%s).  Please check your internet connection and try again.  If this problem persists, you should check Stripe's service status at https://twitter.com/stripestatus, or let us know at support@stripe.com." % (api_base, )
-    elif e[0] == pycurl.E_SSL_CACERT or e[0] == pycurl.E_SSL_PEER_CERTIFICATE:
-      msg = "Could not verify Stripe's SSL certificate.  Please make sure that your network is not intercepting certificates.  (Try going to %s in your browser.)  If this problem persists, let us know at support@stripe.com." % (api_base, )
-    else:
-      msg = "Unexpected error communicating with Stripe.  If this problem persists, let us know at support@stripe.com."
-    msg = textwrap.fill(msg) + "\n\n(Network error: " + e[1] + ")"
-    raise APIConnectionError(msg)
-
-  def urlfetch_request(self, meth, abs_url, headers, params):
-    args = {}
-    if meth == 'post':
-      args['payload'] = self.encode(params)
-    elif meth == 'get' or meth == 'delete':
-      abs_url = self.build_url(abs_url, params)
-    else:
-      raise APIConnectionError('Unrecognized HTTP method %r.  This may indicate a bug in the Stripe bindings.  Please contact support@stripe.com for assistance.' % (meth, ))
-    args['url'] = abs_url
-    args['method'] = meth
-    args['headers'] = headers
-    # Google App Engine doesn't let us specify our own cert bundle.
-    # However, that's ok because the CA bundle they use recognizes
-    # api.stripe.com.
-    args['validate_certificate'] = verify_ssl_certs
-    # GAE requests time out after 60 seconds, so make sure we leave
-    # some time for the application to handle a slow Stripe
-    args['deadline'] = 55
-
-    try:
-      result = urlfetch.fetch(**args)
-    except urlfetch.Error, e:
-      self.handle_urlfetch_error(e, abs_url)
-    return result.content, result.status_code
-
-  def handle_urlfetch_error(self, e, abs_url):
-    if isinstance(e, urlfetch.InvalidURLError):
-      msg = "The Stripe library attempted to fetch an invalid URL (%r).  This is likely due to a bug in the Stripe Python bindings.  Please let us know at support@stripe.com." % (abs_url, )
-    elif isinstance(e, urlfetch.DownloadError):
-      msg = "There were a problem retrieving data from Stripe."
-    elif isinstance(e, urlfetch.ResponseTooLargeError):
-      msg = "There was a problem receiving all of your data from Stripe.  This is likely due to a bug in Stripe.  Please let us know at support@stripe.com."
-    else:
-      msg = "Unexpected error communicating with Stripe.  If this problem persists, let us know at support@stripe.com."
-    msg = textwrap.fill(msg) + "\n\n(Network error: " + str(e) + ")"
-    raise APIConnectionError(msg)
-
-  def urllib2_request(self, meth, abs_url, headers, params):
-    if meth == 'get':
-      abs_url = self.build_url(abs_url, params)
-      req = urllib2.Request(abs_url, None, headers)
-    elif meth == 'post':
-      body = self.encode(params).encode('utf-8')
-      req = urllib2.Request(abs_url, body, headers)
-    elif meth == 'delete':
-      abs_url = self.build_url(abs_url, params)
-      req = urllib2.Request(abs_url, None, headers)
-      req.get_method = lambda: 'DELETE'
-    else:
-      raise APIConnectionError('Unrecognized HTTP method %r.  This may indicate a bug in the Stripe bindings.  Please contact support@stripe.com for assistance.' % (meth, ))
-
-    try:
-      response = urllib2.urlopen(req)
-      rbody = response.read()
-      rcode = response.code
-    except urllib2.HTTPError, e:
-      rcode = e.code
-      rbody = e.read()
-    except (urllib2.URLError, ValueError), e:
-      self.handle_urllib2_error(e, abs_url)
-    return rbody, rcode
-
-  def handle_urllib2_error(self, e, abs_url):
-    msg = "Unexpected error communicating with Stripe.  If this problem persists, let us know at support@stripe.com."
-    msg = textwrap.fill(msg) + "\n\n(Network error: " + str(e) + ")"
-    raise APIConnectionError(msg)
 
 
 class StripeObject(object):
@@ -640,8 +310,8 @@ class APIResource(StripeObject):
   def instance_url(self):
     id = self.get('id')
     if not id:
-      raise InvalidRequestError('Could not determine which URL to request: %s instance has invalid ID: %r' % (type(self).__name__, id), 'id')
-    id = APIRequestor._utf8(id)
+      raise stripe_exceptions.InvalidRequestError('Could not determine which URL to request: %s instance has invalid ID: %r' % (type(self).__name__, id), 'id')
+    id = encoders.utf8(id)
     base = self.class_url()
     extn = urllib.quote_plus(id)
     return "%s/%s" % (base, extn)
@@ -662,7 +332,7 @@ class ListObject(StripeObject):
   def retrieve(self, id, **params):
     requestor = APIRequestor(self.api_key)
     base = self.get('url')
-    id = APIRequestor._utf8(id)
+    id = encoders.utf8(id)
     extn = urllib.quote_plus(id)
     url =  "%s/%s" % (base, extn)
     response, api_key = requestor.request('get', url, params)
